remove SpinCollect and its server. Will be built seperately from MustacheMachine
