+++ src/pages/Home/index.style.module.css
@@ -60,7 +60,7 @@
 	align-items: center;
 	justify-content: center;
 	border-bottom: .1rem solid #555;
-	height: 5.9rem;
+	height: 5.8rem;
 	scroll-snap-align: start;
 	scroll-snap-stop: always;
 }
