--- src/pages/SpinCollect/components/AppMgmt/index.style.css
@@ -1,28 +1,0 @@
-.app-mgmt-container {
-	grid-row: AppMgmt;
-	display: flex;
-	background-color: #7c7b7b;
-}
-
-.mgmt-operations-container {
-	display: flex;
-	flex-direction: column;
-	width: 40rem;
-	height: 100%;
-	background-color: slategray;
-}
-
-.mgmt-operations-container h1 {
-	height: 8rem;
-	background-color: lightslategray;
-}
-
-.mgmt-controls {
-	flex-grow: 1;
-	background-color: steelblue;
-}
-
-.results-container {
-	flex-grow: 1;
-	background-color: blue;
-}
