--- src/pages/SpinCollect/components/Collection/index.style.css
@@ -1,84 +1,0 @@
-.SpinCollect-collection-container {
-	flex-grow: 1;
-	display: flex;
-	flex-direction: column;
-	border: 0.1rem solid black;
-}
-
-.collection-navigation {
-	min-height: 5rem;
-	background-color: #666666;
-}
-
-.spinCollect-spin-card-container {
-	flex-grow: 1;
-	display: grid;
-	grid-template-columns: repeat(4, auto);
-	justify-content: space-evenly;
-	overflow-x: scroll;
-	padding-top: 0.5rem;
-}
-
-.spin-card {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-  justify-content: center;
-	height: 30rem;
-	width: 20rem;
-	font-size: 1.5rem;
-	background-color: aliceblue;
-  border-radius: 0.5rem;
-  margin: 0.5rem auto;
-  padding-top: .5rem;
-}
-
-.no-data-found {
-
-}
-
-.spin-card:hover {
-	transform: scale(1.025);
-	background-color: #c4c4c4;
-}
-
-.card-info {
-	border: 0.10rem solid black;
-	border-radius: 0.25rem;
-	height: 10rem;
-	width: 90%;
-	margin: auto;
-	font-size: 1.5rem;
-	display: flex;
-	flex-direction: column;
-	justify-content: space-evenly;
-	align-items: center;
-	text-align: center;
-}
-
-.spin-card-album-name {
-
-}
-
-.spin-card-album-title {
-	display: flex;
-	align-items: center;
-  justify-content: center;
-  margin: auto;
-}
-
-.spin-card-album-img {
-	width: 90%;
-	border-radius: .5rem;
-}
-
-.album-genres {
-	display: flex;
-	flex-direction: row;
-	width: 100%;
-
-}
-
-.album-genres > div {
-/*	flex-grow: 1;*/
-}
