--- src/pages/SpinCollect/components/Header/index.style.css
@@ -1,41 +1,0 @@
-.spinCollect-header-container {
-	grid-row: Header;
-	display: flex;
-	flex-direction: row;
-	background-color: darkgray;
-}
-
-.spinCollect-header-title {
-	flex-grow: 1;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	font-size: 3.5rem;
-}
-
-.login-status-container {
-  width: 20rem;
-  border-left: .16rem solid black;
-  border-bottom: .16rem solid black;
-}
-
-.is-logged-in-container {
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	height: 100%;
-}
-
-.active-user-name {
-	flex-grow: 1;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	font-size: 3rem;
-}
-
-.is-logged-in-container p  {
-	font-size: 1.2rem;
-	height: 1.5rem;
-}
