--- src/pages/SpinCollect/components/LoginPortal/index.style.css
@@ -1,3 +1,0 @@
-.spinConnect-login-portal-container {
-	
-}
