--- src/pages/SpinCollect/index.style.css
@@ -1,34 +1,0 @@
-:root {
-	--ac-sand: 					#F7DE8D;
-	--ac-orange:				#FBB040;
-	--ac-green: 				#AAC27E;
-	--ac-green-opaque: 	#AAC27E69;
-	--ac-blue:					#9FCFCA;
-	--ac-brown: 				#594A42;
-}
-
-*,
-*:before,
-*:after {
-	margin: 0;
-	box-sizing: inherit;
-}
-
-html {
-	font-size: 62.5%;
-}
-
-html, body {
-	box-sizing: border-box;
-	overflow: auto;
-}
-
-body {
-}
-
-.discogs-app-container {
-	display: grid;
-	grid-template-rows: [Header] 8rem
-											[AppMgmt] calc(100vh - 8rem);
-	background-color: lightgray;
-}
