--- src/pages/SpinCollect/utils/constants.js
@@ -1,10 +1,0 @@
-export const requestTokenURL = `https://api.discogs.com/oauth/request_token`
-export const authorizeURL = `https://www.discogs.com/oauth/authorize`
-export const accesTokenURL = `https://api.discogs.com/oauth/access_token`
-
-export const mustachineMachineRouter = '/api/v1'
-export const spinConnectRouter = `${mustachineMachineRouter}/spin_collect`
-export const loginRoute = `${spinConnectRouter}/user/login_user`
-
-
-const loginUserRoute = '/api/v1/spin_collect/user/login_user'
